requirements.txt
mlx~=0.22.0
mlx-lm~=0.21.1
transformers~=4.48.2
torch~=2.6.0
accelerate~=1.3.0
python-dotenv~=1.0.1